backend/api/requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
astroid==2.5.1
atomicwrites==1.4.0
attrs==19.3.0
cachetools==4.2.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==2.2
django-filter==2.4.0
django-import-export==2.5.0
django-utils-six==2.0
djangorestframework==3.12.3
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.0.1
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.8
importlib-metadata==1.5.0
isort==5.8.0
lazy-object-proxy==1.6.0
MarkupPy==1.14
mccabe==0.6.1
more-itertools==8.2.0
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.1
pep8==1.7.1
Pillow==8.0.0
pluggy==0.13.1
protobuf==3.15.6
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.1
pylint==2.7.2
pyparsing==2.4.6
pyTelegramBotAPI==3.7.6
pytest==5.3.5
pytest-django==3.8.0
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2020.1
PyYAML==5.4.1
requests==2.22.0
rsa==4.7.2
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tablib==3.0.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.6
wcwidth==0.1.8
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==2.2.0


gunicorn==20.0.4
psycopg2-binary
djoser
reportlab
